Fix Turbopack build failure after Next.js 16 upgrade
Dependabot bumped next to 16.1.5 but left @next/mdx at 15.1.3. Next 16
defaults to Turbopack for builds, which broke in two ways:

1. @next/mdx@15 writes to experimental.turbo, which is now invalid
   (Next 16 uses a top-level turbopack key instead)
2. Turbopack serializes loader options, so function references in the
   MDX plugin arrays can't be passed through

Fix by bumping @next/mdx to match, passing plugins as resolvable
strings (mdx-js-loader imports them at runtime), and extracting the
two inline custom plugins to separate .mjs files.

tsconfig.json was auto-updated by Next to set jsx: react-jsx.
